Extract Capabilities Negotiation info from TPM (#1014)
* Extract Capabilities Negotiation info from TPM

+++ keylime/src/structures/capabilities_negotiation.rs
@@ -64,15 +64,15 @@ pub struct Capabilities {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 // Do not serialize the struct name, only the fields
 pub struct ShaValues {
-    pub sha1: Vec<u8>,
-    pub sha256: Vec<u8>,
+    pub sha1: Vec<u32>,
+    pub sha256: Vec<u32>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct CertificationKey {
     pub key_algorithm: String,
     pub key_class: String,
-    pub key_size: u32,
+    pub key_size: usize,
     pub server_identifier: String,
     pub local_identifier: String,
     pub public: String,
